9p: Declare File in PathManager
