+++ src/remote.cc
@@ -674,259 +674,6 @@ void send_command(StringView session, StringView command)
     write(sock, {buffer.data(), buffer.data() + buffer.size()});
 }
 
-class File {
-private:
-    static Glob* p(const char* name)
-    {
-        return new Glob(name);
-    }
-
-    struct Entry {
-        Vector<Glob*> path;
-        FileType* type;
-    };
-
-public:
-    typedef uint32_t Fid;
-
-    enum class Type : uint8_t {
-        DMDIR    = 0x80,
-        DMAPPEND = 0x40,
-        DMEXCL   = 0x20,
-        DMTMP    = 0x04
-    };
-    friend constexpr bool with_bit_ops(Meta::Type<Type>) { return true; }
-
-#pragma pack(push,1)
-    struct Qid {
-        Type     type;
-        uint32_t version;
-        uint64_t path;
-    };
-#pragma pack(pop)
-    static_assert(sizeof(Qid) == 13, "compiler has added padding to Qid");
-
-    File(Vector<String> path, FileType* type)
-      : m_path(path), m_type{type}
-    {}
-
-    Vector<RemoteBuffer> contents() const;
-    std::unique_ptr<File> walk(const String& name) const;
-
-    Type type() const
-    {
-        if (not m_type)
-            return Type::DMDIR;
-        else
-            return Type(0);
-    }
-
-    Vector<String> path() const
-    {
-        return m_path;
-    }
-
-    String fullname() const
-    {
-        if (m_path.empty())
-            return "/";
-        return join(m_path, '/', false);
-    }
-
-    Qid qid() const
-    {
-        String data = fullname();
-        uint64_t path_hash = hash_data(data.data(), size_t(int(data.length())));
-        return { type(), 0, path_hash };
-    }
-
-    uint32_t mode() const
-    {
-        uint32_t mode = uint32_t(type()) << 24;
-        if (type() & Type::DMDIR)
-            mode |= 0755;
-        else
-            mode |= 0644;
-        return mode;
-    }
-
-    uint64_t length() const
-    {
-        if (m_type)
-            return m_type->read(m_path).size();
-        else
-            return 0;
-    }
-
-    String basename() const
-    {
-        if (m_path.empty())
-            return "";
-        return *m_path.rbegin();
-    }
-
-    RemoteBuffer stat() const
-    {
-        RemoteBuffer stat_data;
-        {
-            NinePFieldWriter fields{stat_data};
-            fields.write<uint16_t>(0);  // type, "for kernel use"
-            fields.write<uint32_t>(0);  // dev, "for kernel use"
-            fields.write(qid());
-            fields.write<uint32_t>(mode());
-            fields.write<uint32_t>(0); // atime
-            fields.write<uint32_t>(0); // mtime
-            fields.write<uint64_t>(length());
-            fields.write(basename());
-            fields.write(get_user_name());
-            fields.write("group");
-            fields.write(get_user_name());
-        }
-
-        RemoteBuffer result;
-        {
-            NinePFieldWriter result_fields{result};
-            result_fields.write<uint16_t>(int(stat_data.size()));
-            result_fields.write(stat_data.data(), int(stat_data.size()));
-        }
-        return result;
-    }
-
-private:
-    Vector<String> m_path;
-    FileType* m_type;
-    static Entry m_entries[];
-
-    bool is_our_entry(const Entry& entry) const
-    {
-        if (entry.path.size() != m_path.size() + 1)
-            return false;
-        auto it_a = m_path.begin();
-        auto it_b = entry.path.begin();
-        for (; it_a != m_path.end(); ++it_a, ++it_b)
-            if (not (*it_b)->matches(*it_a))
-                return false;
-        return true;
-    }
-};
-
-RemoteBuffer to_remote_buffer(const char *data)
-{
-    return RemoteBuffer{ data, data + int(strlen(data)) };
-}
-
-RemoteBuffer to_remote_buffer(const StringView& s)
-{
-    return RemoteBuffer{ s.begin(), s.end() };
-}
-
-Context& path_context(const Vector<String>& path)
-{
-    auto it = std::find_if(ClientManager::instance().begin(),
-                           ClientManager::instance().end(),
-                           [&](auto& client) { return client->context().name() == path[1]; });
-    kak_assert(it != ClientManager::instance().end());
-    return (*it)->context();
-}
-
-struct ClientCursorByteOffsetFileType : public FileType
-{
-    RemoteBuffer read(const Vector<String>& path) const
-    {
-        auto& context = path_context(path);
-        auto cursor = context.selections().main().cursor();
-        return to_remote_buffer(to_string(context.buffer().distance({0,0}, cursor)));
-    }
-} client_cursor_byte_offset_file_type{};
-
-struct ClientCursorCharColumnFileType : public FileType
-{
-    RemoteBuffer read(const Vector<String>& path) const
-    {
-        auto& context = path_context(path);
-        auto coord = context.selections().main().cursor();
-        return to_remote_buffer(to_string(context.buffer()[coord.line].char_count_to(coord.column) + 1));
-    }
-} client_cursor_char_column_file_type{};
-
-struct ClientPidFileType : public FileType
-{
-    RemoteBuffer read(const Vector<String>& path) const
-    {
-        auto& context = path_context(path);
-        return to_remote_buffer(format("{}", context.client().pid()));
-    }
-} client_pid_file_type{};
-
-struct NameFileType : public FileType
-{
-    RemoteBuffer read(const Vector<String>& path) const
-    {
-        return to_remote_buffer(Server::instance().session());
-    }
-} name_file_type{};
-
-struct VersionFileType : public FileType
-{
-    RemoteBuffer read(const Vector<String>& path) const
-    {
-        extern const char* version;
-        return to_remote_buffer(version);
-    }
-} version_file_type{};
-
-File::Entry File::m_entries[] = {
-    { {p("clients")},                                             nullptr },
-    { {p("clients"), p("$client_name")},                          nullptr },
-    { {p("clients"), p("$client_name"), p("cursor_byte_offset")}, &client_cursor_byte_offset_file_type },
-    { {p("clients"), p("$client_name"), p("cursor_char_column")}, &client_cursor_char_column_file_type },
-    { {p("clients"), p("$client_name"), p("pid")},                &client_pid_file_type },
-    { {p("name")},                                                &name_file_type },
-    { {p("version")},                                             &version_file_type },
-};
-
-Vector<RemoteBuffer> File::contents() const
-{
-    if (m_type)
-    {
-        Vector<RemoteBuffer> res;
-        res.push_back(m_type->read(m_path));
-        return res;
-    }
-    else
-    {
-        Vector<RemoteBuffer> res;
-        for (const auto& entry : m_entries)
-        {
-            if (not is_our_entry(entry))
-                continue;
-            Vector<String> parts = (*entry.path.rbegin())->expand();
-            for (auto& basename : parts) {
-                Vector<String> path{m_path};
-                path.push_back(std::move(basename));
-                std::unique_ptr<File> file(new File(std::move(path), entry.type));
-                res.push_back(file->stat());
-            }
-        }
-        return res;
-    }
-}
-
-std::unique_ptr<File> File::walk(const String& name) const
-{
-    for (const auto& entry : m_entries)
-    {
-        if (not is_our_entry(entry))
-            continue;
-        if (not (*entry.path.rbegin())->matches(name))
-            continue;
-        Vector<String> path{m_path};
-        path.push_back(name);
-        return std::unique_ptr<File>(new File(std::move(path), entry.type));
-    }
-    return {};
-}
-
 // A client accepter handle a connection until it closes or a nul byte is
 // recieved. Everything recieved before is considered to be a command.
 //
